Update drawterm; hopefully this version doesn't crash.
It had been crashing somewhere in exportfs/exportsrv and I'd
got as far as doing this:

--- a/exportfs/exportsrv.c
+++ b/exportfs/exportsrv.c
@@ -481,6 +481,8 @@ DEBUG(DFD, "blockingslave %d rendez got %p\n", pid, m);
                p = rendezvous((void*)(uintptr)pid, (void*)(uintptr)pid);
                if((uintptr)p == ~(uintptr)0)                   /* Interrupted */
                        continue;
+               if(p == NULL) /* I don't know why this would happen. */
+                       continue;

                DEBUG(DFD, "\tslave: %d %F b %d p %d\n", pid, &p->work, p->busy, p->pid);
                if(p->flushtag != NOTAG)

But that just *froze* it instead of crashing.  So, evidently off.

cinap's apparently rejiggered the doodad and all of that has changed.
So maybe p is never null and it doesn't get locked any more.